pyproject.toml
[project]
name = "project3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.3",
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.8",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
